Update: Change some lines
