+++ Assets/Scenes/TestScenes/Boss.unity
@@ -435,10 +435,6 @@ PrefabInstance:
       propertyPath: <PlayerTransform>k__BackingField
       value: 
       objectReference: {fileID: 1303831235}
-    - target: {fileID: 0}
-      propertyPath: <InvincibilityTime>k__BackingField
-      value: 0.3
-      objectReference: {fileID: 0}
     - target: {fileID: 4930487610698954789, guid: 28a7c46d9c6d0cd49a2365ceacb97701, type: 3}
       propertyPath: m_RootOrder
       value: 0
@@ -487,9 +483,9 @@ PrefabInstance:
       propertyPath: m_Name
       value: MadMantis
       objectReference: {fileID: 0}
-    - target: {fileID: 5126614541931499845, guid: 28a7c46d9c6d0cd49a2365ceacb97701, type: 3}
-      propertyPath: m_TagString
-      value: Boss
+    - target: {fileID: 0}
+      propertyPath: <InvincibilityTime>k__BackingField
+      value: 0.3
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 28a7c46d9c6d0cd49a2365ceacb97701, type: 3}
