+++ Assets/Scripts/Entity/Boss/MadMantis.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c6321e6b6bf141cd9576292edec1dbe0
+timeCreated: 1678307430
