+++ Assets/Scripts/Entity/Boss/MadMantis/EnragedIdle.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace Boss.MadMantis
+{
+    public class EnragedIdle : StateMachineBehaviour
+    {
+        public override void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
+        {
+            animator.GetComponent<Manager>().StartJump();
+            animator.GetComponent<Manager>().StartAttack();
+        }
+    }
+}
