+++ Assets/Scripts/Entity/Boss/MadMantis/EnragedIdle.cs.meta
