+++ Assets/Scripts/Entity/Boss/MadMantis/FlyingdIdle.cs
@@ -1,12 +1,12 @@
 using UnityEngine;
 
-namespace Boss
+namespace Boss.MadMantis
 {
-    public class MadMantisIdleBehavior : StateMachineBehaviour
+    public class FlyingdIdle : StateMachineBehaviour
     {
         public override void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
         {
-            animator.GetComponent<MadMantisManager>().StartAttack();
+            animator.GetComponent<Manager>().Fly();
         }
     }
 }
