+++ Assets/Scripts/Entity/Boss/MadMantis/FlyingdIdle.cs.meta
