+++ Assets/Scripts/Entity/Boss/MadMantis/IdleBehavior.cs.meta
