+++ Assets/Scripts/Entity/Boss/MadMantis/Manager.cs.meta
