+++ Assets/Scripts/Entity/Boss/MadMantis/ShaderManager.cs.meta
