--- Assets/Scripts/Entity/Boss/MadMantisEnrage.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MadMantisEnrage : StateMachineBehaviour
-{
-    public override void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
-    {
-        animator.GetComponent<MadMantisShaderManager>().Enrage();
-    }
-}
