--- Assets/Scripts/Entity/Boss/OldMadMantisHealthController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b8ec52f8c1e8e6f409ca10a5c9ace2a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
